Move message_expiry module to iggy::cmd namespace
