+++ iggy/src/cmd/mod.rs
@@ -1,3 +1,4 @@
 pub mod partitions;
 pub mod streams;
 pub mod topics;
+pub mod utils;
